# TSDB: re-use iterator when moving between series